+++ example/bounce/game.dart
@@ -2,6 +2,7 @@ library bounce;
 
 import 'dart:html';
 import 'package:boarding/pieces.dart';
+import 'package:boarding/util.dart';
 import 'package:boarding/boarding.dart';
 
 part 'view/board.dart';