+++ example/rush/game.dart
@@ -19,7 +19,7 @@ main() {
   final model = new CarParkingModel();
   final grid = new SquareGrid(6);
   final board = new Board(querySelector('#canvas'), grid, model,
-      //withLines: false, area: 'beginner', parking: 1);
-      withLines: false, area: 'intermediate', parking: 2);
+      withLines: false, area: 'beginner', parking: 1);
+      //withLines: false, area: 'intermediate', parking: 2);
   board.draw();
 }