+++ lib/util.dart
@@ -3,4 +3,5 @@ library util;
 import 'dart:math';
 
 part 'util/color.dart';
+part 'util/geo.dart';
 part 'util/random.dart';